+++ K3D/Engine/Source/Scene/Scene.cpp
@@ -19,7 +19,7 @@ K3D::Scene::Scene(std::shared_ptr<RenderingDevice>& renderingDevice) :
 	unsigned int w = 0,h = 0;
 	K3D::Framework::GetInstance().GetWindowSize(w,h);
 	//カメラのイニシャライズ
-	_mainCamera->InitializeCameraFOV(90.0f, w, h,
+	_mainCamera->InitializeCameraFOV(90.0f, static_cast<float>(w),static_cast<float>( h),
 		0.001f, 1000.0f, Vector3(0.0f, 10.0f, -10.0f), Vector3(), Vector3::up);
 	//カメラのデプスステンシルバッファのイニシャライズ
 	_mainCamera->InitializeCameraDepthStencill(DXGI_FORMAT::DXGI_FORMAT_D32_FLOAT, w, h);
